Add gpx support as input
GPX (GPS Exchange Format) is an XML schema designed as a common GPS data format. Users of Strava, Garmin and other apps/software can export their trips in GPX format to then provide to this tool.

Moved the trips to 'outputs' and add 'inputs' to hold sample .gpx files.

Also introduced some functions as the code grows, to keep things managable.

Note that the generated map is still using the static coordinates, TODO to move this to use the gpx data (hopefully not complicated).